requirements.txt
python-mysql-connector
ts3==2.0.0b3